Added a search bar to allows the user to search their own individual workouts. Also changed the name of our app to actually GymSmart instead of "New"
